+++ pyviews/core/__init__.py
@@ -2,7 +2,7 @@
 
 from .error import ViewInfo, PyViewsError, error_handling
 from .binding import Binding, BindingError, BindingCallback
-from .observable import Observable, ObservableEntity, InheritedDict
+from .observable import Bindable, BindableEntity, InheritedDict
 from .reflection import import_path
 from .rendering import Node, InstanceNode, Setter
 from .xml import XmlAttr, XmlNode, XmlError